+++ UniExplore/base/views/__init__.py
@@ -1,0 +1,11 @@
+from .home import *
+from .sso import *
+from .challanges_and_responses import *
+from .leaderboard import *
+from .likes import *
+from .lockout import *
+from .login import *
+from .password_reset import *
+from .profile import *
+from .recent_activity import *
+from .register import *
